Rewrote the scheduler to use consumers
Each instruction is now aware if it has been scheduled or not. If the set of
producers is empty then the target instruction is a viable target for
scheduling. There is still the issue of removing duplicate references of a
target register in a given instruction but this is a start.

Anyway, each time an instruction is scheduled, the consumers of the target
are informed that the target instruction has been completed.